refactor: ConnectionHandler 패키지 분리
